+++ NEWS.rst
@@ -19,6 +19,10 @@ pkgdev 0.2.18 (unreleased)
 - bugs: fix ``--filter-stablereqs`` collapsing a target which matches several
   packages down to a single one (Arthur Zamarin)
 
+- bugs: under ``--filter-stablereqs``, when pulling dependency, prefer the
+  version the stablereq check considers due for stabilization if possible
+  (Arthur Zamarin)
+
 pkgdev 0.2.17 (2026-08-14)
 --------------------------